requirements.txt
numpy
matplotlib
seaborn
requests
pandas
python-dotenv
pymongo
plotly
pyspark
streamlit
folium
scipy
statsmodels
scikit-learn
